fix all mistakes
