+++ js/main.js
@@ -1,13 +1,12 @@
 import {renderGallery} from './gallery.js';
-import {formSubmit} from './form.js';
 import {getData} from './api.js';
 import {showAlert, debounce} from './universal.js';
 import {loadLocalFile} from './local-picture.js';
 import {init, getFilter} from './filter.js';
+import {formSubmit} from './form.js';
 
 loadLocalFile();
 formSubmit();
-
 
 getData()
   .then((data) => {
